Write tests for user summary
I think this sorts out most of the JUnit tests for the
server side of the user summary example. Now I need to
actually write the code that makes this pass, which is
mostly implementing `UserController.getUserSummary()`.

I've also done *some* refactoring as I've written
this, but there is some housekeeping that ought to be
done. Things in the test like `getCompany()` are too
much like `getName()`, for example.