Implement user summary server-side code
This basically completes at least one version of the server-side code
for issue #23, providing an example of computing summary data from the
database.

This uses MongoDB's map-reduce tool, and it's not _too_ bad except for
all the insane JSON/BSON document manipulation that's necessary to do
this kind of stuff in Java. Ugh.

There's still no client-side code for this, and no write-up or
documentation anywhere. There probably should be some substantial
comments added to the code as well.