Improve logging on Travis for failing JUnit tests
The current Gradle setup provides almost no meaningful info if a JUnit
test fails on the server side code. Apparently we can set the
`testLogging` field in Gradle to increase the amount of info we get,
as per https://mrhaki.blogspot.com/2013/05/gradle-goodness-show-more-information.html

Hopefully this will improve the info we get in Travis-CI when tests
fail (& improve the info we get when we run tests locally as well).