Fix broken E2E tests
I extended the set of the users we seed the database with, and that
broke almost all the E2E tests because they were focussed on particular
individuals in the set of users.

This substantially revises the E2E tests for the `user-list` component
so that (a) they pass again and (b) they are hopefully less fragile
to future change.

This took a *really* long time, in large part because I got really
lost in how to get values out of the DOM.
https://stackoverflow.com/a/46545517/2557372 saved the day by showing
us how to use `getAttribute('textContent')` to get the content we
needed.

Big thanks to @kklamberty for working on this with me for several hours
today while I was hopefully confused.