Start implementing the client-side for company summaries
This is a start on implementing the client-side for company summaries,
but there still a lot of work to be done.

One thing I'm a little confused about is how to best handle JSON we
get back from the server. I'd been quite happy with the idea of using
the company as the key that maps to the bucket data as the value. I'm
not sure how to parse that, though, since that's essentially a "raw"
JSON map and that doesn't seem to work well with all the nifty
auto-parsing stuff that Angular would otherwise provide.