Upgrade Angular (CLI)
The goal here was to upgrade Angular CLI to avoid a bunch of security
warnings. I ended up having to upgrade quite a lot to get things to
work, and I think we've effectively upgraded to Angular 5 at this point.

The tests all seem to pass, so I think we're good. The one remaining
issue is that some of the Gradle tasks don't work because of changes
to the flags for `ng`.