Reduce compatibility warnings
This eliminates a few library compatibility
warnings caused by `flex-layout`. We still have
two which it looks like we're stuck with for now.
I don't think that `@angular/flex-layout` is
getting a lot of development attention, so we
may have to live with some of these problems if
we're going to continue to use it.