Roll node back to v8
I can't get Travis to deal with versions of
node higher than 8, so I'm rolling back to that
in the hopes that this will allow our builds to
succeed. I think we could use `nvm` on Travis to
specify a more recent version, but hopefully
this will get us over the current hump.