Update `Server.java` to work on the droplet
The setup of the Java Spark server needed to be updated to work
on the droplet. This was based on various previous projects and
assembled with the valuable assistance of @floogulinc and
@ahnafy.

The key thing here was removing the reroute to `localhost`
and adding a new route that properly finds the resources
in `public`.

I'm not sure if this will run correctly in development mode
in the lab, so we'll need to check that.